+++ flake.nix
@@ -1,30 +1,45 @@
 {
   inputs = {
-    nixpkgs.url = "github:nixos/nixpkgs/nixpkgs-unstable";
+    nixpkgs.url = "github:nixos/nixpkgs/nixos-25.05";
     flake-utils.url = "github:numtide/flake-utils";
+    nixpkgs-unstable.url = "github:nixos/nixpkgs/nixos-unstable";
   };
 
-  outputs = { self, nixpkgs, flake-utils, }:
-    flake-utils.lib.eachDefaultSystem (system:
+  outputs =
+    {
+      self,
+      nixpkgs,
+      nixpkgs-unstable,
+      flake-utils,
+    }:
+    flake-utils.lib.eachDefaultSystem (
+      system:
       let
-        pkgs = nixpkgs.legacyPackages.${system};
-
-        corepackEnable = pkgs.runCommand "corepack-enable" { } ''
+        stable = nixpkgs.legacyPackages.${system};
+        unstable = nixpkgs-unstable.legacyPackages.${system};
+        nodejs = stable.nodejs_22;
+        corepackEnable = stable.runCommand "corepack-enable" { } ''
           mkdir -p $out/bin
-          ${pkgs.nodejs_20}/bin/corepack enable --install-directory $out/bin
+          ${nodejs}/bin/corepack enable --install-directory $out/bin
         '';
       in
       {
         devShells = {
-          default = with pkgs;
+          default =
+            with stable;
             mkShell {
+              BIOME_BINARY = "${unstable.biome}/bin/biome";
               buildInputs = [
-                act
-                nodejs_20
+                # see https://github.com/biomejs/biome-vscode/issues/295
+                # nixos can't deal with statically linked binaries, so
+                # we need to use the biome nix package
+                unstable.biome
+                nodejs
                 sqlite
                 corepackEnable
               ];
             };
         };
-      });
+      }
+    );
 }