+++ prettier.config.mjs
@@ -1,7 +1,0 @@
-export default {
-    semi: true,
-    singleQuote: true,
-    trailingComma: 'all',
-    bracketSpacing: false,
-    arrowParens: 'avoid',
-};