+++ workspace/tons-of-tests/src/internal/diff.ts
@@ -2,7 +2,7 @@ import colors from 'ansi-colors';
 
 import type {DisplayConfig} from '../DisplayConfig.js';
 import type {DiffColumn, DiffColumnNames, DiffContext} from '../Show.js';
-import type {Diff, TestRunResults} from '../Test.js';
+import type {Diff} from '../Test.js';
 import {
     colorNegative,
     colorPositive,
@@ -70,7 +70,7 @@ export const showDiff = ({
     const cfg = {...makeDefault(), ...displayConfig};
     const _columns = columns.filter(c => selectColumns.includes(c.name));
 
-    const display = {
+    const _display = {
         precision: diff.precision.toFixed(2),
         recall: diff.recall.toFixed(2),
     };