+++ workspace/tons-of-tests/src/internal/format-diff.spec.ts
@@ -19,10 +19,10 @@ t.describe('format-diff', () => {
         {a: undefined, b: 1, expected: '1'},
         {a: 1, b: undefined, expected: '1'},
         {a: 1, b: 2, expected: '1 => 2'},
-        {a: [], b: [1], expected: `[\n  0: 1\n]`},
-        {a: [1], b: [], expected: `[\n  0: 1\n]`},
-        {a: {foo: undefined}, b: {foo: 1}, expected: `{\n  foo: 1\n}`},
-        {a: {foo: 1}, b: {foo: undefined}, expected: `{\n  foo: 1\n}`},
+        {a: [], b: [1], expected: '[\n  0: 1\n]'},
+        {a: [1], b: [], expected: '[\n  0: 1\n]'},
+        {a: {foo: undefined}, b: {foo: 1}, expected: '{\n  foo: 1\n}'},
+        {a: {foo: 1}, b: {foo: undefined}, expected: '{\n  foo: 1\n}'},
     ])('non-empty diff', ({a, b, expected}) => {
         const result = c.unstyle(formatDiff(diff(a, b)) ?? '');
         t.expect(result).toEqual(expected);